# [NG] Datagrid: rendering performance improvements