--- a/src/clarity-angular/data/datagrid/render/constants.ts
+++ b/src/clarity-angular/data/datagrid/render/constants.ts
@@ -4,5 +4,6 @@
  * The full license information can be found in LICENSE in the root directory of this project.
  */
 
+export const NO_LAYOUT_CLASS = "datagrid-no-layout";
 export const COMPUTE_WIDTH_CLASS = "datagrid-computing-columns-width";
 export const STRICT_WIDTH_CLASS = "datagrid-fixed-width";