--- a/src/clarity-angular/data/datagrid/render/table-renderer.spec.ts
+++ b/src/clarity-angular/data/datagrid/render/table-renderer.spec.ts
@@ -7,18 +7,28 @@ import {Component} from "@angular/core";
 
 import {TestContext} from "../helpers.spec";
 
+import {MOCK_DOM_ADAPTER_PROVIDER} from "./dom-adapter.mock";
 import {DatagridRenderOrganizer} from "./render-organizer";
 import {MOCK_ORGANIZER_PROVIDER, MockDatagridRenderOrganizer} from "./render-organizer.mock";
 import {DatagridTableRenderer} from "./table-renderer";
 
 export default function(): void {
-    describe("DatagridTableRenderer directive", function() {
+    describe("DatagridTableRenderer component", function() {
         let context: TestContext<DatagridTableRenderer, SimpleTest>;
         let organizer: MockDatagridRenderOrganizer;
 
         beforeEach(function() {
-            context = this.create(DatagridTableRenderer, SimpleTest, [MOCK_ORGANIZER_PROVIDER]);
+            context =
+                this.create(DatagridTableRenderer, SimpleTest, [MOCK_ORGANIZER_PROVIDER, MOCK_DOM_ADAPTER_PROVIDER]);
             organizer = <MockDatagridRenderOrganizer>context.getClarityProvider(DatagridRenderOrganizer);
+        });
+
+        it("toggles in and out of no layout mode when notified", function() {
+            expect(context.clarityElement.classList).not.toContain("datagrid-no-layout");
+            organizer.noLayout.next(true);
+            expect(context.clarityElement.classList).toContain("datagrid-no-layout");
+            organizer.noLayout.next(false);
+            expect(context.clarityElement.classList).not.toContain("datagrid-no-layout");
         });
 
         it("toggles in and out of table mode when notified", function() {
@@ -28,8 +38,29 @@ export default function(): void {
             organizer.tableMode.next(false);
             expect(context.clarityElement.classList).not.toContain("datagrid-computing-columns-width");
         });
+
+        it("moves the header in and out of the body when notified", function() {
+            const body = context.clarityElement.querySelector(".datagrid-body");
+            expect(context.clarityElement.textContent).toMatch("Hello");
+            expect(body.textContent).not.toMatch("Hello");
+            expect(body.textContent).toMatch("World");
+            organizer.tableMode.next(true);
+            expect(body.textContent).toMatch("Hello");
+            expect(body.textContent).toMatch("World");
+            organizer.tableMode.next(false);
+            expect(context.clarityElement.textContent).toMatch("Hello");
+            expect(body.textContent).not.toMatch("Hello");
+            expect(body.textContent).toMatch("World");
+        });
     });
 }
 
-@Component({template: `<div clrDgTableWrapper>Hello World</div>`})
+@Component({
+    template: `
+        <clr-dg-table-wrapper>
+            <div ngProjectAs="[clrDgHead]">Hello</div>
+            World
+        </clr-dg-table-wrapper>
+    `
+})
 class SimpleTest {}