… src/clarity-angular/data/datagrid/render/body-renderer.ts
… src/clarity-angular/data/datagrid/render/head-renderer.ts
… src/clarity-angular/data/datagrid/render/row-master-renderer.spec.ts
… src/clarity-angular/data/datagrid/render/row-master-renderer.ts