Some browsers (I'm frowning at Firefox right now) are really slow
rendering flexboxes, so we had to manually modify some internal
rendering logic to improve the rendering speed of a new datagrid on
these browsers.

Trues to fix #1572, but large datagrids without pagination (or large pages) will still render slowly, if only because Angular runs change detection on each cell. We can try and improve this in the future, but this is a first pass that addresses specific flexbox issues.

Results for initial rendering of 1000 items per page on a new datagrid:

Browser | Before | After
------- | ------ | -----
Chrome | 1917 ms | 1500 ms
Firefox | 2546 ms | 1540 ms
Safari | 1690 ms | 1150 ms
Edge* | ∞ s | 41s
IE11* | ∞ s | 52s

* Windows browsers were tested on a VM with limited resources, so these numbers should not be used to compare browsers, only the before/after improvement relative to that specific browser.